project refactor
Code cleanup in an effort to make things a bit more uniform accross the server as well as to add documentation and make things a bit easier to navigate

+++ habbgo/config/config.go
@@ -9,7 +9,7 @@ import (
 type Config struct {
 	Server struct {
 		Host     string `yaml:"host"`
-		Port     int16  `yaml:"port"`
+		Port     int    `yaml:"port"`
 		MaxConns int    `yaml:"maxconns"`
 	}
 	Log struct {
